Fix compiler warnings (signed/unsigned, unused vars, Boost noise)
This update resolves multiple compiler warnings across the C++ codebase:

- Signed/unsigned comparison fixes (hsize_t, Eigen::Index, R_xlen_t)
- Initialization of potentially uninitialized variables
- Removal/silencing of unused variables
- Added -DBOOST_NO_CONCEPT_CHECKING to reduce Boost/BH warnings
- No functional changes; behavior remains identical

Improves CRAN readiness and ensures cleaner builds on all platforms.

+++ inst/include/hdf5Algebra/matrixQR.hpp
@@ -186,7 +186,9 @@ inline void RcppQRHdf5( BigDataStatMeth::hdf5Dataset* dsA,
         irank = lu_decomp.rank();
         
         
-        if (irank == count[0] + 1 || irank == count[1] + 1 )
+        //..//if (irank == count[0] + 1 || irank == count[1] + 1 )
+        if (static_cast<unsigned long long>(irank) == count[0] + 1ULL ||
+            static_cast<unsigned long long>(irank) == count[1] + 1ULL)
         {
             Eigen::MatrixXd R = qr.matrixQR().template triangularView<Eigen::Upper>();
             dsR->createDataset( R.rows(), R.cols(), "real" );
